+++ qa/suites/orch/cephadm/smb/tasks/deploy_smb_mgr_basic.yaml
@@ -14,7 +14,7 @@ overrides:
     log-only-match:
       - CEPHADM_
 tasks:
-- cephadm.configure_samba_client_container:
+- smb.configure_samba_client_container:
     role: host.b
 - cephadm:
     single_host_defaults: true
