+++ qa/suites/orch/cephadm/smb/tasks/deploy_smb_mgr_res_ports.yaml
@@ -14,7 +14,7 @@ overrides:
     log-only-match:
       - CEPHADM_
 tasks:
-- cephadm.deploy_samba_ad_dc:
+- smb.deploy_samba_ad_dc:
     role: host.b
 - cephadm:
     single_host_defaults: true
